added slope movement, and updated ground checking
added slope movement (i did some mathematics to calculate the distance between feet and ground in the case that the radius for feet is 0.3 spanning from the center of capsule also i added the calculation for distance between the feet and an acceptable slope to check whether they are in the right distance already, if they are, isGrounded became true. used a formula for finding radius when slope degree is known. also i found out that ontriggerstay is called on physics frame (aka fixedupdate) meaning that i can make ground checking more responsive by turning isGround false at every end of physics frame.