Spike trap added, also made onSlopeGroundCheck more versatile.
slope ground check now also qualify for 0 degree (non existential slope, only activate on those that does not use groundTag to avoid double setting of the isGrounded bool), but it is still recommended to use groundTag for ground in order to get a more accurate result. Also added spike trap.